Fix memoization
The ORM selectors did not take advantage of reselect memoization due
to how they were written. Fix all the selectors to such that memoization
is scoped to the Entity type.

A few breaking changes:
- recordType is set in the model definition. e.g. new Shot({}, 'shot')
- Selectors no longer reinstantiate the entities and just fetch the
records from the store. It is up to the reducer to store the entities
as ORM records properly.
- pagination() and order() no longer accept any parameter and always
return an immutable object